Move download-related code out of `dist` and into its own `download` mod
The idea here is to reduce the API surface of `dist`, while
consolidating download-related code into a single place. Was also able
to remove a number of pointless pass-arounds of the
`notification_handler` attached to DownloadCfg (haven't caught them all
yet).
