+++ src/rustup-dist/src/lib.rs
@@ -35,5 +35,6 @@ pub mod notifications;
 pub mod prefix;
 pub mod component;
 pub mod manifestation;
+pub mod download;
 pub mod manifest;
 pub mod config;
