+++ src/rustup-dist/tests/dist.rs
@@ -18,7 +18,7 @@ use rustup_dist::prefix::InstallPrefix;
 use rustup_dist::ErrorKind;
 use rustup_dist::errors::Result;
 use rustup_dist::dist::{ToolchainDesc, TargetTriple, DEFAULT_DIST_SERVER};
-use rustup_dist::dist;
+use rustup_dist::download::DownloadCfg;
 use rustup_dist::Notification;
 use rustup_utils::utils;
 use rustup_utils::raw as utils_raw;
@@ -269,9 +269,8 @@ fn update_from_dist(dist_server: &Url,
                     prefix: &InstallPrefix,
                     add: &[Component],
                     remove: &[Component],
-                    download_cfg: &dist::DownloadCfg,
-                    temp_cfg: &temp::Cfg,
-                    notify_handler: &Fn(Notification)) -> Result<UpdateStatus> {
+                    download_cfg: &DownloadCfg,
+                    temp_cfg: &temp::Cfg) -> Result<UpdateStatus> {
 
     // Download the dist manifest and place it into the installation prefix
     let ref manifest_url = try!(make_manifest_url(dist_server, toolchain));
@@ -289,7 +288,7 @@ fn update_from_dist(dist_server: &Url,
         remove_extensions: remove.to_owned(),
     };
 
-    manifestation.update(&manifest, changes, download_cfg, notify_handler.clone())
+    manifestation.update(&manifest, changes, download_cfg, download_cfg.notify_handler.clone())
 }
 
 fn make_manifest_url(dist_server: &Url, toolchain: &ToolchainDesc) -> Result<Url> {
@@ -309,7 +308,7 @@ fn uninstall(toolchain: &ToolchainDesc, prefix: &InstallPrefix, temp_cfg: &temp:
 }
 
 fn setup(edit: Option<&Fn(&str, &mut MockPackage)>,
-         f: &Fn(&Url, &ToolchainDesc, &InstallPrefix, &dist::DownloadCfg, &temp::Cfg)) {
+         f: &Fn(&Url, &ToolchainDesc, &InstallPrefix, &DownloadCfg, &temp::Cfg)) {
     let dist_tempdir = TempDir::new("multirust").unwrap();
     create_mock_dist_server(dist_tempdir.path(), edit).write(&[ManifestVersion::V2]);
 
@@ -323,7 +322,7 @@ fn setup(edit: Option<&Fn(&str, &mut MockPackage)>,
     let ref url = Url::parse(&format!("file://{}", dist_tempdir.path().to_string_lossy())).unwrap();
     let ref toolchain = ToolchainDesc::from_str("nightly-x86_64-apple-darwin").unwrap();
     let ref prefix = InstallPrefix::from(prefix_tempdir.path().to_owned());
-    let ref download_cfg = dist::DownloadCfg {
+    let ref download_cfg = DownloadCfg {
         dist_root: "phony",
         temp_cfg: temp_cfg,
         download_dir: &prefix.path().to_owned().join("downloads"),
@@ -336,7 +335,7 @@ fn setup(edit: Option<&Fn(&str, &mut MockPackage)>,
 #[test]
 fn initial_install() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
 
         assert!(utils::path_exists(&prefix.path().join("bin/rustc")));
         assert!(utils::path_exists(&prefix.path().join("lib/libstd.rlib")));
@@ -346,7 +345,7 @@ fn initial_install() {
 #[test]
 fn test_uninstall() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         uninstall(toolchain, prefix, temp_cfg, &|_| ()).unwrap();
 
         assert!(!utils::path_exists(&prefix.path().join("bin/rustc")));
@@ -357,7 +356,7 @@ fn test_uninstall() {
 #[test]
 fn uninstall_removes_config_file() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         assert!(utils::path_exists(&prefix.manifest_file("multirust-config.toml")));
         uninstall(toolchain, prefix, temp_cfg, &|_| ()).unwrap();
         assert!(!utils::path_exists(&prefix.manifest_file("multirust-config.toml")));
@@ -368,10 +367,10 @@ fn uninstall_removes_config_file() {
 fn upgrade() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
         change_channel_date(url, "nightly", "2016-02-01");
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         assert_eq!("2016-02-01", utils_raw::read_file(&prefix.path().join("bin/rustc")).unwrap());
         change_channel_date(url, "nightly", "2016-02-02");
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         assert_eq!("2016-02-02", utils_raw::read_file(&prefix.path().join("bin/rustc")).unwrap());
     });
 }
@@ -390,10 +389,10 @@ fn update_removes_components_that_dont_exist() {
     };
     setup(Some(edit), &|url, toolchain, prefix, download_cfg, temp_cfg| {
         change_channel_date(url, "nightly", "2016-02-01");
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         assert!(utils::path_exists(&prefix.path().join("bin/bonus")));
         change_channel_date(url, "nightly", "2016-02-02");
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         assert!(!utils::path_exists(&prefix.path().join("bin/bonus")));
     });
 }
@@ -411,13 +410,13 @@ fn update_preserves_extensions() {
             ];
 
         change_channel_date(url, "nightly", "2016-02-01");
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
 
         assert!(utils::path_exists(&prefix.path().join("lib/i686-apple-darwin/libstd.rlib")));
         assert!(utils::path_exists(&prefix.path().join("lib/i686-unknown-linux-gnu/libstd.rlib")));
 
         change_channel_date(url, "nightly", "2016-02-02");
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
 
         assert!(utils::path_exists(&prefix.path().join("lib/i686-apple-darwin/libstd.rlib")));
         assert!(utils::path_exists(&prefix.path().join("lib/i686-unknown-linux-gnu/libstd.rlib")));
@@ -450,12 +449,12 @@ fn update_preserves_extensions_that_became_components() {
             ];
 
         change_channel_date(url, "nightly", "2016-02-01");
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
 
         assert!(utils::path_exists(&prefix.path().join("bin/bonus")));
 
         change_channel_date(url, "nightly", "2016-02-02");
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         assert!(utils::path_exists(&prefix.path().join("bin/bonus")));
     });
 }
@@ -480,10 +479,10 @@ fn update_preserves_components_that_became_extensions() {
     };
     setup(Some(edit), &|url, toolchain, prefix, download_cfg, temp_cfg| {
         change_channel_date(url, "nightly", "2016-02-01");
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         assert!(utils::path_exists(&prefix.path().join("bin/bonus")));
         change_channel_date(url, "nightly", "2016-02-02");
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         assert!(utils::path_exists(&prefix.path().join("bin/bonus")));
     });
 }
@@ -491,9 +490,9 @@ fn update_preserves_components_that_became_extensions() {
 #[test]
 fn update_makes_no_changes_for_identical_manifest() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
-        let status = update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        let status = update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         assert_eq!(status, UpdateStatus::Changed);
-        let status = update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        let status = update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
         assert_eq!(status, UpdateStatus::Unchanged);
     });
 }
@@ -510,7 +509,7 @@ fn add_extensions_for_initial_install() {
             }
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
         assert!(utils::path_exists(&prefix.path().join("lib/i686-apple-darwin/libstd.rlib")));
         assert!(utils::path_exists(&prefix.path().join("lib/i686-unknown-linux-gnu/libstd.rlib")));
     });
@@ -519,7 +518,7 @@ fn add_extensions_for_initial_install() {
 #[test]
 fn add_extensions_for_same_manifest() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
 
         let ref adds = vec![
             Component {
@@ -530,7 +529,7 @@ fn add_extensions_for_same_manifest() {
             }
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
 
         assert!(utils::path_exists(&prefix.path().join("lib/i686-apple-darwin/libstd.rlib")));
         assert!(utils::path_exists(&prefix.path().join("lib/i686-unknown-linux-gnu/libstd.rlib")));
@@ -542,7 +541,7 @@ fn add_extensions_for_upgrade() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
         change_channel_date(url, "nightly", "2016-02-01");
 
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
 
         change_channel_date(url, "nightly", "2016-02-02");
 
@@ -555,7 +554,7 @@ fn add_extensions_for_upgrade() {
             }
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
 
         assert!(utils::path_exists(&prefix.path().join("lib/i686-apple-darwin/libstd.rlib")));
         assert!(utils::path_exists(&prefix.path().join("lib/i686-unknown-linux-gnu/libstd.rlib")));
@@ -572,7 +571,7 @@ fn add_extension_not_in_manifest() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
     });
 }
 
@@ -586,7 +585,7 @@ fn add_extension_that_is_required_component() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
     });
 }
 
@@ -603,7 +602,7 @@ fn add_extensions_for_same_manifest_when_extension_already_installed() {
 #[test]
 fn add_extensions_does_not_remove_other_components() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
 
         let ref adds = vec![
             Component {
@@ -611,7 +610,7 @@ fn add_extensions_does_not_remove_other_components() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
 
         assert!(utils::path_exists(&prefix.path().join("bin/rustc")));
     });
@@ -628,7 +627,7 @@ fn remove_extensions_for_initial_install() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg).unwrap();
     });
 }
 
@@ -644,7 +643,7 @@ fn remove_extensions_for_same_manifest() {
             }
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
 
         let ref removes = vec![
             Component {
@@ -652,7 +651,7 @@ fn remove_extensions_for_same_manifest() {
             },
             ];
 
-        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg).unwrap();
 
         assert!(!utils::path_exists(&prefix.path().join("lib/i686-apple-darwin/libstd.rlib")));
         assert!(utils::path_exists(&prefix.path().join("lib/i686-unknown-linux-gnu/libstd.rlib")));
@@ -673,7 +672,7 @@ fn remove_extensions_for_upgrade() {
             }
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
 
         change_channel_date(url, "nightly", "2016-02-02");
 
@@ -683,7 +682,7 @@ fn remove_extensions_for_upgrade() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg).unwrap();
 
         assert!(!utils::path_exists(&prefix.path().join("lib/i686-apple-darwin/libstd.rlib")));
         assert!(utils::path_exists(&prefix.path().join("lib/i686-unknown-linux-gnu/libstd.rlib")));
@@ -696,7 +695,7 @@ fn remove_extension_not_in_manifest() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
         change_channel_date(url, "nightly", "2016-02-01");
 
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
 
         change_channel_date(url, "nightly", "2016-02-02");
 
@@ -706,7 +705,7 @@ fn remove_extension_not_in_manifest() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg).unwrap();
     });
 }
 
@@ -733,7 +732,7 @@ fn remove_extension_not_in_manifest_but_is_already_installed() {
                 pkg: "bonus".to_string(), target: Some(TargetTriple::from_str("x86_64-apple-darwin"))
             },
         ];
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
         assert!(utils::path_exists(&prefix.path().join("bin/bonus")));
 
         change_channel_date(url, "nightly", "2016-02-02");
@@ -743,7 +742,7 @@ fn remove_extension_not_in_manifest_but_is_already_installed() {
                 pkg: "bonus".to_string(), target: Some(TargetTriple::from_str("x86_64-apple-darwin"))
             },
         ];
-        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg).unwrap();
     });
 }
 
@@ -751,7 +750,7 @@ fn remove_extension_not_in_manifest_but_is_already_installed() {
 #[should_panic]
 fn remove_extension_that_is_required_component() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
 
         let ref removes = vec![
             Component {
@@ -759,7 +758,7 @@ fn remove_extension_that_is_required_component() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg).unwrap();
     });
 }
 
@@ -767,7 +766,7 @@ fn remove_extension_that_is_required_component() {
 #[should_panic]
 fn remove_extension_not_installed() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
 
         let ref removes = vec![
             Component {
@@ -775,7 +774,7 @@ fn remove_extension_not_installed() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg).unwrap();
     });
 }
 
@@ -793,7 +792,7 @@ fn remove_extensions_does_not_remove_other_components() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
 
         let ref removes = vec![
             Component {
@@ -801,7 +800,7 @@ fn remove_extensions_does_not_remove_other_components() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], removes, download_cfg, temp_cfg).unwrap();
 
         assert!(utils::path_exists(&prefix.path().join("bin/rustc")));
     });
@@ -818,7 +817,7 @@ fn add_and_remove_for_upgrade() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
 
         change_channel_date(url, "nightly", "2016-02-02");
 
@@ -834,7 +833,7 @@ fn add_and_remove_for_upgrade() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, removes, download_cfg, temp_cfg).unwrap();
 
         assert!(utils::path_exists(&prefix.path().join("lib/i686-apple-darwin/libstd.rlib")));
         assert!(!utils::path_exists(&prefix.path().join("lib/i686-unknown-linux-gnu/libstd.rlib")));
@@ -850,7 +849,7 @@ fn add_and_remove() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, &[], download_cfg, temp_cfg).unwrap();
 
         let ref adds = vec![
             Component {
@@ -864,7 +863,7 @@ fn add_and_remove() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, removes, download_cfg, temp_cfg).unwrap();
 
         assert!(utils::path_exists(&prefix.path().join("lib/i686-apple-darwin/libstd.rlib")));
         assert!(!utils::path_exists(&prefix.path().join("lib/i686-unknown-linux-gnu/libstd.rlib")));
@@ -875,7 +874,7 @@ fn add_and_remove() {
 #[should_panic]
 fn add_and_remove_same_component() {
     setup(None, &|url, toolchain, prefix, download_cfg, temp_cfg| {
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap();
 
         let ref adds = vec![
             Component {
@@ -889,7 +888,7 @@ fn add_and_remove_same_component() {
             },
         ];
 
-        update_from_dist(url, toolchain, prefix, adds, removes, download_cfg, temp_cfg, &|_| ()).unwrap();
+        update_from_dist(url, toolchain, prefix, adds, removes, download_cfg, temp_cfg).unwrap();
     });
 }
 
@@ -900,7 +899,7 @@ fn bad_component_hash() {
         let path = path.join("dist/2016-02-02/rustc-nightly-x86_64-apple-darwin.tar.gz");
         utils_raw::write_file(&path, "bogus").unwrap();
 
-        let err = update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap_err();
+        let err = update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap_err();
 
         match *err.kind() {
             ErrorKind::ComponentDownloadFailed(_) => (),
@@ -916,7 +915,7 @@ fn unable_to_download_component() {
         let path = path.join("dist/2016-02-02/rustc-nightly-x86_64-apple-darwin.tar.gz");
         fs::remove_file(&path).unwrap();
 
-        let err = update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| ()).unwrap_err();
+        let err = update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg).unwrap_err();
 
         match *err.kind() {
             ErrorKind::ComponentDownloadFailed(..) => (),
@@ -942,16 +941,25 @@ fn reuse_downloaded_file() {
 
         prevent_installation(prefix);
 
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|_| {}).unwrap_err();
+        let reuse_notification_fired = Arc::new(Cell::new(false));
+
+        let download_cfg = DownloadCfg {
+            dist_root: download_cfg.dist_root,
+            temp_cfg: download_cfg.temp_cfg,
+            download_dir: download_cfg.download_dir,
+            notify_handler: &|n| {
+                if let Notification::FileAlreadyDownloaded = n {
+                    reuse_notification_fired.set(true);
+                }
+            }
+        };
+
+        update_from_dist(url, toolchain, prefix, &[], &[], &download_cfg, temp_cfg).unwrap_err();
+        assert!(!reuse_notification_fired.get());
 
         allow_installation(&prefix);
 
-        let reuse_notification_fired = Arc::new(Cell::new(false));
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|n| {
-            if let Notification::FileAlreadyDownloaded = n {
-                reuse_notification_fired.set(true);
-            }
-        }).unwrap();
+        update_from_dist(url, toolchain, prefix, &[], &[], &download_cfg, temp_cfg).unwrap();
 
         assert!(reuse_notification_fired.get());
     })
@@ -969,12 +977,18 @@ fn checks_files_hashes_before_reuse() {
         println!("wrote previous download to {}", prev_download.display());
 
         let noticed_bad_checksum = Arc::new(Cell::new(false));
-        update_from_dist(url, toolchain, prefix, &[], &[], download_cfg, temp_cfg, &|n| {
-            println!("{:?}", n);
-            if let Notification::CachedFileChecksumFailed = n {
-                noticed_bad_checksum.set(true);
+        let download_cfg = DownloadCfg {
+            dist_root: download_cfg.dist_root,
+            temp_cfg: download_cfg.temp_cfg,
+            download_dir: download_cfg.download_dir,
+            notify_handler: &|n| {
+                if let Notification::CachedFileChecksumFailed = n {
+                    noticed_bad_checksum.set(true);
+                }
             }
-        }).unwrap();
+        };
+
+        update_from_dist(url, toolchain, prefix, &[], &[], &download_cfg, temp_cfg).unwrap();
 
         assert!(noticed_bad_checksum.get());
     })
